backend/pyproject.toml
[project]
name = "vorja-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.5",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "gunicorn>=26.0.0",
    "pillow>=12.2.0",
    "stripe>=15.1.0",
]
